refactor: mark gateway as an experimental package
