+++ exp/gateway/internal/models/helpers.go
